+++ src/File/Content.php
@@ -8,12 +8,10 @@ use Innmind\Filesystem\File\Content\{
     Line,
 };
 use Innmind\IO\{
-    IO,
     Streams\Stream,
     Files\Read,
     Stream\Size,
 };
-use Innmind\Url\Path;
 use Innmind\Immutable\{
     Str,
     Sequence,
@@ -28,16 +26,6 @@ final class Content
 {
     private function __construct(private Implementation $implementation)
     {
-    }
-
-    /**
-     * @psalm-pure
-     */
-    public static function atPath(
-        IO $io,
-        Path $path,
-    ): self {
-        return new self(Content\AtPath::of($io, $path));
     }
 
     /**