+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - `Innmind\Filesystem\Adapter\InMemory::new()`
 - `Innmind\Filesystem\Adapter\Filesystem::withCaseSensitivity()`, case sensitivity can be specified as the second argument of `::mount()`
 - `Innmind\Filesystem\Exception\*`
+- `Innmind\Filesystem\File\Content::atPath()`, use `Content::io()` instead
 
 ## 8.1.0 - 2025-05-09